requirements.txt
flask>=2.0.1
pytesseract>=0.3.8
opencv-python-headless>=4.8.0
Pillow>=9.0.0
numpy>=1.24.0 